+++ metricbeat/docs/modules/vsphere.asciidoc
@@ -11,7 +11,7 @@ This file is generated! See scripts/mage/docs_collector.go
 
 The vSphere module uses the https://github.com/vmware/govmomi[Govmomi] library to collect metrics from any Vmware SDK URL (ESXi/VCenter). This library is built for and tested against ESXi and vCenter 5.5, 6.0 and 6.5.
 
-By default it enables the metricsets `datastore`, `datastore_cluster`, `host`, `network`, `resourcepool` and `virtualmachine`.
+By default it enables the metricsets `datastore`, `datastorecluster`, `host`, `network`, `resourcepool` and `virtualmachine`.
 
 [float]
 === Dashboard
@@ -35,7 +35,7 @@ in <<configuration-metricbeat>>. Here is an example configuration:
 metricbeat.modules:
 - module: vsphere
   enabled: true
-  metricsets: ["datastore", "datastore_cluster", "host", "network", "resourcepool", "virtualmachine"]
+  metricsets: ["datastore", "datastorecluster", "host", "network", "resourcepool", "virtualmachine"]
   # Real-time data collection – An ESXi Server collects data for each performance counter every 20 seconds.
   period: 20s
   hosts: ["https://localhost/sdk"]
@@ -55,7 +55,7 @@ The following metricsets are available:
 
 * <<metricbeat-metricset-vsphere-datastore,datastore>>
 
-* <<metricbeat-metricset-vsphere-datastore_cluster,datastore_cluster>>
+* <<metricbeat-metricset-vsphere-datastorecluster,datastorecluster>>
 
 * <<metricbeat-metricset-vsphere-host,host>>
 
@@ -67,7 +67,7 @@ The following metricsets are available:
 
 include::vsphere/datastore.asciidoc[]
 
-include::vsphere/datastore_cluster.asciidoc[]
+include::vsphere/datastorecluster.asciidoc[]
 
 include::vsphere/host.asciidoc[]
