--- metricbeat/docs/modules/vsphere/datastore_cluster.asciidoc
@@ -1,28 +1,0 @@
-////
-This file is generated! See scripts/mage/docs_collector.go
-////
-:edit_url: https://github.com/elastic/beats/edit/main/metricbeat/module/vsphere/datastore_cluster/_meta/docs.asciidoc
-
-
-[[metricbeat-metricset-vsphere-datastore_cluster]]
-=== vSphere datastore_cluster metricset
-
-beta[]
-
-include::../../../module/vsphere/datastore_cluster/_meta/docs.asciidoc[]
-
-
-:edit_url:
-
-==== Fields
-
-For a description of each field in the metricset, see the
-<<exported-fields-vsphere,exported fields>> section.
-
-Here is an example document generated by this metricset:
-
-[source,json]
-----
-include::../../../module/vsphere/datastore_cluster/_meta/data.json[]
-----
-:edit_url!:
