+++ metricbeat/docs/modules/vsphere/datastorecluster.asciidoc
@@ -1,0 +1,29 @@
+////
+This file is generated! See scripts/mage/docs_collector.go
+////
+:edit_url: https://github.com/elastic/beats/edit/main/metricbeat/module/vsphere/datastorecluster/_meta/docs.asciidoc
+
+
+[[metricbeat-metricset-vsphere-datastorecluster]]
+=== vSphere datastorecluster metricset
+
+beta[]
+
+include::../../../module/vsphere/datastorecluster/_meta/docs.asciidoc[]
+
+This is a default metricset. If the host module is unconfigured, this metricset is enabled by default.
+
+:edit_url:
+
+==== Fields
+
+For a description of each field in the metricset, see the
+<<exported-fields-vsphere,exported fields>> section.
+
+Here is an example document generated by this metricset:
+
+[source,json]
+----
+include::../../../module/vsphere/datastorecluster/_meta/data.json[]
+----
+:edit_url!:
