+++ metricbeat/docs/modules_list.asciidoc
@@ -312,7 +312,7 @@ This file is generated! See scripts/mage/docs_collector.go
 .1+| .1+|  |<<metricbeat-metricset-uwsgi-status,status>>   
 |<<metricbeat-module-vsphere,vSphere>>     |image:./images/icon-yes.png[Prebuilt dashboards are available]    |  
 .6+| .6+|  |<<metricbeat-metricset-vsphere-datastore,datastore>>   
-|<<metricbeat-metricset-vsphere-datastore_cluster,datastore_cluster>> beta[]  
+|<<metricbeat-metricset-vsphere-datastorecluster,datastorecluster>> beta[]  
 |<<metricbeat-metricset-vsphere-host,host>>   
 |<<metricbeat-metricset-vsphere-network,network>> beta[]  
 |<<metricbeat-metricset-vsphere-resourcepool,resourcepool>> beta[]
