+++ metricbeat/module/vsphere/_meta/README.md
@@ -1,7 +1,7 @@
 # Testing using GOVCSIM.
 
 
-To test the vsphere module without a real Vmware SDK URL you can use the following setup. Govcsim is a vCenter Server and ESXi API based simulator written using govmomi. It creates a vCenter Server model with a datacenter, hosts, cluster, resource pools, networks and a datastore.
+To test the vsphere module without a real Vmware SDK URL you can use the following setup. Govcsim is a vCenter Server and ESXi API based simulator written using govmomi. It creates a vCenter Server model with a datacenter, datastore cluster, hosts, cluster, resource pools, networks and a datastore.
 
 
 Requirements:
@@ -29,7 +29,7 @@ Now setup your metricbeat config to connect to Govcsim:
 - module: vsphere
   metricsets:
     - datastore
-    - datastore_cluster
+    - datastorecluster
     - host
     - network
     - resourcepool
