+++ metricbeat/module/vsphere/_meta/config.reference.yml
@@ -1,6 +1,6 @@
 - module: vsphere
   enabled: true
-  metricsets: ["datastore", "datastore_cluster", "host", "network", "resourcepool", "virtualmachine"]
+  metricsets: ["datastore", "datastorecluster", "host", "network", "resourcepool", "virtualmachine"]
   # Real-time data collection – An ESXi Server collects data for each performance counter every 20 seconds.
   period: 20s
   hosts: ["https://localhost/sdk"]
