+++ metricbeat/module/vsphere/_meta/config.yml
@@ -1,7 +1,7 @@
 - module: vsphere
   # metricsets:
   #  - datastore
-  #  - datastore_cluster
+  #  - datastorecluster
   #  - host
   #  - network
   #  - resourcepool
