+++ metricbeat/module/vsphere/_meta/docs.asciidoc
@@ -1,6 +1,6 @@
 The vSphere module uses the https://github.com/vmware/govmomi[Govmomi] library to collect metrics from any Vmware SDK URL (ESXi/VCenter). This library is built for and tested against ESXi and vCenter 5.5, 6.0 and 6.5.
 
-By default it enables the metricsets `datastore`, `datastore_cluster`, `host`, `network`, `resourcepool` and `virtualmachine`.
+By default it enables the metricsets `datastore`, `datastorecluster`, `host`, `network`, `resourcepool` and `virtualmachine`.
 
 [float]
 === Dashboard
