--- metricbeat/module/vsphere/datastore_cluster/_meta/docs.asciidoc
@@ -1,1 +1,0 @@
-This is the datastorecluster metricset of the module vsphere.
