+++ metricbeat/module/vsphere/datastorecluster/_meta/docs.asciidoc
@@ -1,0 +1,1 @@
+This is the Datastore Cluster metricset of the module vsphere.
