+++ metricbeat/module/vsphere/datastorecluster/datastorecluster_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package datastore_cluster
+package datastorecluster
 
 import (
 	"testing"
