+++ metricbeat/module/vsphere/fields.go
@@ -32,5 +32,5 @@ func init() {
 // AssetVsphere returns asset data.
 // This is the base64 encoded zlib format compressed contents of module/vsphere.
 func AssetVsphere() string {
-	return "eJzUXEtv4zgSvudXFOay3UDac89hgdk0drqBTfegnck1oKmyxQ1FasmSDefXL0jqZVuSZZtyEh8aDT+qvnoXi6V8gRfc3sHa5ikavAEgQRLv4Lf13L/z2w1AgpYbkZPQ6g7+eQMAUH4KmU4K6X5mUCKzeAcrdgOwFCgTe+e/+gUUy7DNwr1om7svG13k5TsdXHYJtYkljJglXZPrJtlLtvyog8iuHNVrH0YbCmc544K2s6VBnC22hHbnexUwqdVq74MBbO71b4MIniDoJVCKDeLZ3peX2mSM7uCQ/QFM0sRkVJyPjmJ8oIXFJCrOvy0mE8HMOXWCtJxJTJ6XUrP9LxwB+xcajorYCh3YGmjNFzzfHth5+HEn8KV10DrRvuB2o01yopcKiXZrCTNPZdbJNNWWZlwXqltNp9vyR5Et0DjVONJ2gKv7b7cHnSXuf4Qlx5ZJ6V3I8fDsejAIncfy3zlp4/zh++8/4V4rMloCW60MrhhhAt9VXtDvPwvKC4KfORrmCFnnSTBHrlXSjdD9G08/X2tPdXS7ORpkcSP7FyMfJo6wUCsfLrA0Ors0yj1SyQgVrzJnFgv0H2v05mSZiwufkkSGDhQwzxh0ZcQOWUAoyISUwnrT9nifJUZFRO+fe3pH0mdd8bOJQv7poUfcdTZ1uD89DAX7xgiK2wNUnu0o155N+lK/DkDfwrE95zM9u78NfOaysITmaDtY93cLJDauTWwy2v0Bk6HGMG5efUzxEMlAjq37k/idXt2C+DQgLNiyMOU6uXXW8yxPd0rXQD/bnPG4IRRAO+LgiYNWF8NudxiXHEG+ue5BqMCkDofqdIVqLYxWGSqaXXZGyYvQp2bpa8xm+v6vv53iHr699jhhXlTJJRrjYM4xnP2BLB5jfxw7xrfJyRMXomSnz+qpRw2aaQpxTd+GoArtcA8WYV9mrVMTW0UO87RdcarYrnMVL4xBRXILC3SFlGtli8wfCEGbCj4qcuhOTl1euATXgmNVVFmogpOXVUFA7AWt6wq4znKJhMAUzO/n390bGVNJSCt5urWCMwkB6LgO0ks2UaPwTaxStAQleVgzWSAwbrS13tEdc+tPus5M9XFrPPDoJ43KCKqOgTBQ8P1608sI+wLIeAq24+A12qHit5N98B0nQlV3lnHwxx81zYvM5x2Hz6vcOXdoKQ1zgrRAX9CJZJhpM9VE78ETvxzdVIO8WPimmd9djC5uW/6tGv/0zFaQNtq8PEduBn4EskOlv+Q8dRuixiOZpgUpqY9oQCocC6aSjUgonZFhymYu801TJFxKAkawSQVPw9F9wyy02EJSGNeNOORCEZo1kzN4dKcTg7lBi4qs/7RGXQ1fKnFOD4ADPRjkKNbXVULF831oYLJKVcVH2+auZtXih7JVpbNObZwtX874C5Ld8fNporDk1BZzOAgraLXrTYurZjMKFBqjzTXVthGUQuB6ugpLtFfSZBvqeVr1oRYTZGhb1BDUcQizQpLgzNIVTF/zOj92GrgT2/4Q62mWb+l1cuMfYB0HMTE6zzG5guHblaBSZ8X9JKgTG72uUBeBvF60D8KLfQ/2mCJo1+BICSkySWnJwvfG9ZTClfEjw9wGY5GT6DmanDcN87ovye7iKgcnPkOeCvOt79gmu1VcC0MFk5AxngrVN9Lcc/ibfQQRLpzK090X42eJSft+oK2SoYE/4xytFQvZ7U8LrSUydZqmvqtEcN+tblKkFI3r6p1wBFpVkzkLXKulWBUGEyANudFrkWC4cOls2OuJvf/dLHakHqJ2Ni0XVlJmIUFC7tvyGnq4ihHWFgPrLJ0Q963fZyfYG0jsC3xM6BGCu9cPR7o6LvlFGSejCvL6WV+fVWDHMofhBkMhNxbd7g7PWGjTDHKcqgYPllOVkj9RuTpSlZESwo8hCHEXuB63+b7k8Alnq9ltheLTnJhKmEk+38JXYcmIRUGYPIWEmWtDn/ty9YeLkoMq8C4D5hyUzU6T1YXhmGstpyhfv0r64BicfLl95Bo7mj+FC8jD22GIYKd/MSu4vyrO0XjxFUefOIQlwe1t5xVyG12CGVNd1/XvA56/JpWY7e+YtvF1GQUGDNOm3y33UdlHyg8Ht8VOF1VYVNLfAqWMQiw9PbjOppTahdgtMAucSV6EFm2xha+/5ocKg73CIXiPTWPLFpjtCNa2WrmMtZdGupN4uM6JG3iTeM2qQEudQ9xdHpdr+E/HqVQMFCSkeA0Zbi+IOq+oapF657j7gvkTyDXk8rdZk4p1rQxyDW3t5pCgtkY/0fLHKZ5S5phrSV9mmdJjxieaWF6UG7FmPSsSR2UdIaeT0bW3Lgb6TSwsrAxTZfN1IGxoe5RWX2zK3LE4baKsS/xxogdaH0Hylrj1r0pNLJA2iOpwofskTWyYH/x9AFVUx/B/WI867KOeL/mCSam1mlh2K17r82HJERIj1micDk6M7XGC6TWaFFnXee7jloYmP36qNPQ5+M5GSAkLrBfdDr2HLfQaQZANTU5D8X+FMD7RRmgzqpXIq5WPRkOVwTuiq3dxkzSYwjng08NlVYTrLDdo+9aVjko9Mo4aaRuGlSEbIdt7qU8P5wbTdM8gqNaczLRP+h/34iXilOoNJj/lLcWALeDKY7QTIdZmCD8rf3V0MjWAqJfS6U9O+KOX2NfIpQuDouea0nNz/8RfUmzH7cCOXFS+5YAYHsoq1r8nqSOmiH2u5aOwagXzcOsz8NDUxM/JpG/2nMwQ58jPyfyxZkKyhcSjzNuLwv3DmxjrwnpZzmzibDX3DmQiYS0fC4uyIB5fsTtr4vE0690wPtr2zn0EsLywpLPnUDM6IerFf/Hgz0GEN58vSGv3nnFZrN7btnnXn1KBD3Hf1zxHx6zVXLD2HdpuI/GWF36tJ+1OxNm/sAIfwkD1gv+7ts+5KKc6J/3cPSMxleyck0Zd/ERfSGtW0co1r2YVreMh/5v/BwAA//8QuYyT"
+	return "eJzUXEtv4zgSvudXFOay3UDac89hgdk0drqBTfegnck1oKmyxQ1FasmSDefXL0jqZVuSZZtyEh8aDT+qvnoXi6V8gRfc3sHa5ikavAEgQRLv4Lf13L/z2w1AgpYbkZPQ6g7+eQMAUH4KmU4K6X5mUCKzeAcrdgOwFCgTe+e/+gUUy7DNwr1om7svG13k5TsdXHYJtYkljJglXZPrJtlLtvyog8iuHNVrH0YbCmc544K2s6VBnC22hHbnexUwqdVq74MBbO71b4MIniDoJVCKDeLZ3peX2mSM7uCQ/QFM0sRkVJyPjmJ8oIXFJCrOvy0mE8HMOXWCtJxJTJ6XUrP9LxwB+xcajorYCh3YGmjNFzzfHth5+HEn8KV10DrRvuB2o01yopcKiXZrCTNPZdbJNNWWZlwXqltNp9vyR5Et0DjVONJ2gKv7b7cHnSXuf4Qlx5ZJ6V3I8fDsejAIncfy3zlp4/zh++8/4V4rMloCW60MrhhhAt9VXtDvPwvKC4KfORrmCFnnSTBHrlXSjdD9G08/X2tPdXS7ORpkcSP7FyMfJo6wUCsfLrA0Ors0yj1SyQgVrzJnFgv0H2v05mSZiwufkkSGDhQwzxh0ZcQOWUAoyISUwnrT9nifJUZFRO+fe3pH0mdd8bOJQv7poUfcdTZ1uD89DAX7xgiK2wNUnu0o155N+lK/DkDfwrE95zM9u78N5LKwhOZoN1i3dwskNq5LbBLa/QGTob4wblp9TPEQyUCKrduT+I1e3YH4LCAs2LIu5Tq5dcbzLE/3Sdc/P9uc8bgRFEA74uCJg1YXw243GJecQL655kGowKSOhupwhWotjFYZKppddkTJi9CmZulrzF76/q+/neIevr32OGFeVLklGuNgzjGc/XksHmN/GjvGt0nJE9ehZKfN6ilHDZpp6nBN34agCt1wDxZhX2atQxNbRQ7ztF1wqtiucxUvjEFFcgsLdHWUa2WLzJ8HQZsKPipy6E5OXV64BNeCY1VTWSiCk1dVQUDsBa1rCrjOcomEwBTM7+ff3RsZU0lIK3m6tYIzCQHouAbSSzZRn/BNrFK0BCV5WDNZIDButLXe0R1z6w+6zkz1aWs88OgHjcoIqo6BME/w7XrTygj7Ash4Crbj3DXaoeJ3k33wHSdCVTeWcfDHnzTNi8znHYfPq9w5d+goDXOCtEBf0IlkmGkz1UDvwRO/HN1Uc7xY+KYZ312MLm5b/q2a/vSMVpA22rw8R24GfgSyQ6W/5Dx1G6LGI5mmBSmpj2hAKhwLppKNSCidkWHKZi7zTVMkXEoCRrBJBU/DyX3DLLTYQlIY14045EIRmjWTM3h0pxODuUGLiqz/tEZdzV4qcU4PgAM9GOQo1tdVQsXzfWhgskpVxUfb5q5m1eKHslWls05tnC1fzvgLkt3x82misOTUFnM4CCtotetNi6tmMwoUGqPNNdW2EZRC4Hq6Cku0V9JkG+p5WvWhFhNkaFvUENRxCLNCkuDM0hVMX/M6P3YauBPb/hDraZZv6XVy4x9gHQcxMTrPMbmC4duVoFJnxf0kqBMbva5QF4G8XrQPwot9DfaYImjX4EgJKTJJacnC98b1lMKV8SPD3AZjkZPoOZqcNw3zui/J7uIqByc+Q54K862v2Ca7VFwLQwWTkDGeCtU30txz+Jt9BBEunMrT3RfjZ4lJ+36grZKhgT/jHK0VC9ntTwutJTJ1mqa+q0Rw361uUqQUjevqnXAEWlWTOQtcq6VYFQYTIA250WuRYLhw6WzY64m9/90sdqQeonY2LfdVUmYhQULu2/IaeriKEdYWA9ssnRD3rd9nJ9gbSOwLfEzoEYK71w9Hujou+T0ZJ6MK8vpZX59VYMcyh+EGQyE3Ft3uCs9YaNMMcpyqBg+WU5WSP1G5OlKVkRLCjyEIcfe3Hrf5vuTwCWer2W2F4tOcmEqYST7fwldhyYhFQZg8hYSZa0Of+3L1h4uSgyrwLgPmHJTNSpPVheGYay2nKF+/SvrgGJx8uX3kGjuaP4ULyMPbYYhgp38xK7i/Ks7RePEVR584hCXB7W3nFXIbXYIZU13X9e8Dnr8mlZjtr5i28XUZBQYM06bfLfdR2UfKDwe3xU4XVVhU0t8CpYxCLD09uM6mlNqF2C0wC5xJXoQWbbGFr7/mhwqDvcIheI9NY8sWmO0I1rZauYu1l0a6k3i4zokbeJN4zapAS51D3F0el2v4T8epVAwUJKR4DRluL4g6r6hqkXrnuPuC+RPINeTyt1mTinWtDHINbe3mkKC2Rj/R8scpnlLmmGtJX2aZ0mPGJ5pYXpQbsWY9KxJHZR0hp5PRtbcuBvpNLCysDFNl83UgbGh7lFZfbMrcsThtoqxL/HGiB1ofQfKWuPWvSk0skDaI6nCf+yRNbJgf/H0AVVTH8H9Yjzrso54v+YJJqbWaWHYrXuvzYckREiPWaJwOToztcYLpNZoUWdd57uOWhiY/fqo09Dn4zkZICQusF90OvYct9BpBkA1NTkPxf4UwPtFGaDOqlcirlY9GQ5XBO6Krd3GTNJjCOeDTw2VVhOssN2j71pWOSj0yjhppG4aVIRsh23upTw/nBtN0zyCo1pzMtE/6H/fiJeKU6g0mP+UtxYAt4MpjtBMh1mYIPyt/dXQyNYCol9LpT074o5fY18ilC4Oi55rSc3P/xF9SbMftwI5cVL7lgBgeyirWvyepI6aIfa7lk7BqBfNw6zPw0NTEz8mkb/aczBDnyM/J/LFmQrKFxKPM24vC/cObGOvCelnObOJsNfcOZCJhLR8Li7IgHl+xO2vi8TTr3TA+2vbOfQSwvLCks+dQMzoh6sV/8eCvQYQ3ny9Ia/eecVms3tu2eddfUoEPcd/XPEfHrNVcsPYd2m4j8ZYXfq0n7U7E2b+wAh/CQPWC/7u2z7kopzon/dw9IzGV7JyTRl38RF9Ia1bRyjWvZhWt4xn/m/8HAAD//4wUjDQ="
 }
